refactor: move relevant code to setup and algebra module
