+++ labrador/src/lib.rs
@@ -1,2 +1,4 @@
+pub mod setup;
 pub mod prover;
 pub mod verifier;
+pub mod algebra;
